backend/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.4
distro==1.9.0
filelock==3.13.3
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
fsspec==2024.3.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.14
langchain-community==0.0.31
langchain-core==0.1.40
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
langsmith==0.1.40
Mako==1.3.2
MarkupSafe==2.1.5
marshmallow==3.21.1
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
mysql-connector-python==8.3.0
networkx==3.2.1
numpy==1.26.4
openai==1.16.1
orjson==3.10.0
packaging==23.2
pillow==10.3.0
pinecone-client==3.2.2
pydantic==2.6.4
pydantic_core==2.16.3
pypdf==4.1.0
PyPDF2==3.0.1
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.13.0
sentence-transformers==2.6.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.4.0
tiktoken==0.6.0
tokenizers==0.15.2
torch==2.2.2
tqdm==4.66.2
transformers==4.39.3
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
uuid==1.30
Werkzeug==3.0.2
wikipedia==1.4.0
yarl==1.9.4
